clinics and doctorClinics services finished

+++ src/doctor-clinics.module/doctor-clinics.service.ts
@@ -10,13 +10,38 @@ import {DoctorClinics} from "./doctor-clinics.model";
 import {DoctorClinicsDto} from "../dto/doctor-clinics/doctor-clinics.dto";
 import {TQuerySettings} from "../types";
 import {InjectModel} from "@nestjs/sequelize";
+import {Doctor} from "../doctors.module/doctor.model";
+import {Clinic} from "../clinics.module/clinic.model";
 
 @Injectable()
-export class DoctorClinicsService implements ICreateService<DoctorClinics, DoctorClinicsDto>, IGetByIdService<DoctorClinics>, IGetAllService<DoctorClinics>, IUpdateService<DoctorClinics, DoctorClinicsDto>, IDeleteService {
-    constructor(@InjectModel(DoctorClinics) private doctorClinicsRepository: typeof DoctorClinics) {
+export class DoctorClinicsService
+    implements ICreateService<DoctorClinics, DoctorClinicsDto>,
+        IGetByIdService<DoctorClinics>,
+        IGetAllService<DoctorClinics>,
+        IUpdateService<DoctorClinics, DoctorClinicsDto>,
+        IDeleteService {
+    constructor(@InjectModel(DoctorClinics) private doctorClinicsRepository: typeof DoctorClinics,
+                @InjectModel(Doctor) private doctorRepository: typeof Doctor,
+                @InjectModel(Clinic) private clinicRepository: typeof Clinic) {
     }
 
     async create(dto: DoctorClinicsDto): Promise<DoctorClinics> {
+        let doctor = await this.doctorRepository.findOne({
+            where: {
+                id: dto.doctor_id
+            }
+        });
+
+        if (!doctor) throw new HttpException("Doctor not found", 404);
+
+        let clinic = await this.doctorRepository.findOne({
+            where: {
+                id: dto.clinic_id
+            }
+        })
+
+        if (!clinic) throw new HttpException("Clinic not found", 404);
+
         let doctorClinicRecord = await this.doctorClinicsRepository.findOne({
             where: {
                 doctor_id: dto.doctor_id,
@@ -34,7 +59,54 @@ export class DoctorClinicsService implements ICreateService<DoctorClinics, Docto
 
         if (!doctorClinic) throw new HttpException("Doctor clinic record was not found", 404);
 
+        let doctor = await this.doctorRepository.findOne({
+            where: {
+                id: doctorClinic.doctor_id
+            }
+        });
+
+        if (!doctor) throw new HttpException("Doctor not found", 404);
+
+        let clinic = await this.doctorRepository.findOne({
+            where: {
+                id: doctorClinic.clinic_id
+            }
+        })
+
+        if (!clinic) throw new HttpException("Clinic not found", 404);
+
         await this.doctorClinicsRepository.destroy({where: {id: id}});
+
+        return null;
+    }
+
+    async deleteByDoctorAndClinicId(doctorId: number, clinicId: number) {
+        let doctorClinic = await this.doctorClinicsRepository.findOne({
+            where: {
+                clinic_id: clinicId,
+                doctor_id: doctorId
+            }
+        });
+
+        if (!doctorClinic) throw new HttpException("Doctor clinic record was not found", 404);
+
+        let doctor = await this.doctorRepository.findOne({
+            where: {
+                id: doctorClinic.doctor_id
+            }
+        });
+
+        if (!doctor) throw new HttpException("Doctor not found", 404);
+
+        let clinic = await this.doctorRepository.findOne({
+            where: {
+                id: doctorClinic.clinic_id
+            }
+        })
+
+        if (!clinic) throw new HttpException("Clinic not found", 404);
+
+        await this.doctorClinicsRepository.destroy({where: {id: doctorClinic.id}});
 
         return null;
     }
@@ -59,7 +131,7 @@ export class DoctorClinicsService implements ICreateService<DoctorClinics, Docto
     }
 
     async getById(id: number): Promise<DoctorClinics> {
-        let doctorClinic = await this.doctorClinicsRepository.findOne({ where: { id } });
+        let doctorClinic = await this.doctorClinicsRepository.findOne({where: {id}});
 
         if (!doctorClinic) throw new HttpException("Doctor clinic record not found", 404);
 
@@ -67,7 +139,7 @@ export class DoctorClinicsService implements ICreateService<DoctorClinics, Docto
     }
 
     async update(id: number, updateModel: DoctorClinicsDto): Promise<DoctorClinics> {
-        let doctorClinic = await this.doctorClinicsRepository.findOne({ where: { id } });
+        let doctorClinic = await this.doctorClinicsRepository.findOne({where: {id}});
 
         if (!doctorClinic) throw new HttpException("Doctor clinic not found", 404);
